requirements.txt
Django==2.2.10
psycopg2-binary==2.8.4
django-bootstrap4==1.1.1
mozilla-django-oidc==1.2.3
rules==2.2
django-photologue==3.11
